refactor(AdditionResult): 더 큰수를 계산하도록 하기 위해 long 타입으로 형변환 하도록 변경

+++ src/main/java/calculator/dto/AdditionResult.java
@@ -3,7 +3,7 @@ package calculator.dto;
 public record AdditionResult(String result) {
     public static AdditionResult from(double result) {
         if (result % 1 == 0) {
-            return new AdditionResult(String.valueOf((int) result));
+            return new AdditionResult(String.valueOf((long) result));
         }
         return new AdditionResult(String.valueOf(result));
     }
